+++ lib/queued-graphql/clients/qs-observer/q_s_observer.d.ts
@@ -1,16 +1,12 @@
 import { DocumentNode } from 'graphql';
-import { ClassType } from '../../../helpers';
 import { RumbleshipContext } from '../../../app/rumbleship-context';
+import { ClassType } from '../../../helpers';
 import { QueuedSubscriptionMessage } from '../../servers';
 /**
  * QsrObservers are similar to Resolvers, however rahter than doing a class level
  * decorator, we make all QsrObservers derive from this base calss.. this may change in future
  * to match the style of type-graphql...
  */
-export declare abstract class QueuedSubscriptionObserver {
-    ctx: RumbleshipContext;
-    constructor(ctx: RumbleshipContext);
-}
 export declare const QUEUED_SUBSCRIPTION_OBSERVER_META: unique symbol;
 export interface QueuedSubscriptionObserverMetadata {
     gql_document: DocumentNode | string;
@@ -21,8 +17,8 @@ export interface QueuedSubscriptionObserverMetadata {
 }
 export interface QueuedSubscriptionHandler {
     qso_metadata: QueuedSubscriptionObserverMetadata;
-    observer_class?: ClassType<QueuedSubscriptionObserver>;
-    handler(this: QueuedSubscriptionObserver, message: QueuedSubscriptionMessage): Promise<void>;
+    observer_class?: ClassType<Record<string, any>>;
+    handler(this: Record<string, any>, ctx: RumbleshipContext, message: QueuedSubscriptionMessage): Promise<void>;
 }
-export declare function SubscriptionHandler<T>(metadata: QueuedSubscriptionObserverMetadata): MethodDecorator;
-export declare function getQsoHandlers(target: ClassType<QueuedSubscriptionObserver>): QueuedSubscriptionHandler[];
+export declare function QSObserver<T>(metadata: QueuedSubscriptionObserverMetadata): MethodDecorator;
+export declare function getQsoHandlers(target: ClassType<Record<string, any>>): QueuedSubscriptionHandler[];