+++ src/app/server/init-server.ts
@@ -27,14 +27,13 @@ import hapiRequireHttps = require('hapi-require-https');
 import hapiRequestIdHeader = require('hapi-request-id-header');
 
 import { Oid } from '@rumbleship/oid';
-import { QueuedCacheScopeAndDb, QueuedSubscriptionObserver } from '../../queued-graphql';
+import { QueuedCacheScopeAndDb } from '../../queued-graphql';
 import { addErrorToTraceContext } from '../honeycomb-helpers/add_error_to_trace_context';
 import {
   initQueuedGraphql,
   startQueuedGraphQl,
   stopQueuedGraphQl
 } from '../../queued-graphql/init_queued_graphql';
-import { ClassType } from '../../helpers';
 
 export let globalGraphQlSchema: GraphQLSchema | undefined;
 
@@ -45,7 +44,7 @@ export async function initServer(
   injected_apollo_server_options: Pick<Config, 'plugins' | 'uploads'> = {},
   injected_models: DbModelAndOidScope[],
   injected_schema_options: Omit<BuildSchemaOptions, 'authChecker' | 'pubSub' | 'container'>,
-  injected_subscription_observers: ClassType<QueuedSubscriptionObserver>[],
+
   injected_routes: Hapi.ServerRoute[] = [],
   onContainer: (context: RumbleshipContext, ServiceFactories: ServiceFactoryMap) => void,
   onInitialized: (server: Hapi.Server, pubSub: RfiPubSub) => Promise<void> = (
@@ -251,7 +250,7 @@ export async function initServer(
   // Set up subscriptions for websocket clients
   apolloServer.installSubscriptionHandlers(server.listener);
   // setup subscription server for GooglePubSub  clients
-  initQueuedGraphql(config, globalGraphQlSchema, injected_subscription_observers);
+  initQueuedGraphql(config, globalGraphQlSchema, injected_schema_options.resolvers as any);
 
   server.events.on('start', async () => {
     const ctx = RumbleshipContext.make(__filename, {