+++ src/gql/resolvers/index.ts
@@ -1,5 +1,6 @@
 export * from './base-resolver';
 export * from './base-resolver.interface';
+export * from './base-subscription-resolver.mixin';
 export * from './create-node-notification';
 export * from './node-resolver';
 export * from './relay-resolver.interface';