+++ src/queued-graphql/clients/qs-observer/queued_subscription_observer_manager.ts
@@ -7,14 +7,10 @@ import { ClassType } from '../../../helpers';
 import { RumbleshipContext } from '../../../app/rumbleship-context';
 import { QueuedSubscriptionMessage } from '../../servers';
 import { RfiPubSubSubscription } from '../../shared';
-import { QueuedGqlRequestClientOneInstanceResponder } from '../gql-request/queued-gql-request-client';
+import { QueuedGqlRequestClientSingleInstanceResponder } from '../gql-request/queued-gql-request-client';
 import { syncQsrGql, syncQsrVariables } from './sync_qsr.interface';
-import {
-  getQsoHandlers,
-  QueuedSubscriptionHandler,
-  QueuedSubscriptionObserver
-} from './queued_subscription_observer';
 import { forcePublicProjectPubsub } from '../../../helpers/pubsub-auth-project';
+import { getQsoHandlers, QueuedSubscriptionHandler } from './q_s_observer';
 
 /**
  * Each service has its own pubsub topic that subscription responses are sent to. We subscribe to this
@@ -30,10 +26,10 @@ export class QueuedSubscriptionObserverManager {
   qsrSubscriptionName: string;
   handlers: Map<string, QueuedSubscriptionHandler> = new Map();
   _initialized = false;
-  queuedGqlRequestClient: QueuedGqlRequestClientOneInstanceResponder;
+  queuedGqlRequestClient: QueuedGqlRequestClientSingleInstanceResponder;
   constructor(
     public config: ISharedSchema,
-    protected observers: ClassType<QueuedSubscriptionObserver>[]
+    protected observers: readonly ClassType<Record<string, any>>[]
   ) {
     const pubsub = new GooglePubSub(forcePublicProjectPubsub(this.config.Gcp.Auth));
     this.qsrTopicName = `${this.config.PubSub.topicPrefix}_QSR_PUBLISH_TO.${config.serviceName}`;
@@ -49,9 +45,9 @@ export class QueuedSubscriptionObserverManager {
       this.qsrSubscriptionName,
       false
     );
-    this.queuedGqlRequestClient = new QueuedGqlRequestClientOneInstanceResponder(config);
+    this.queuedGqlRequestClient = new QueuedGqlRequestClientSingleInstanceResponder(config);
   }
-  setHandlers(observers: ClassType<QueuedSubscriptionObserver>[]): void {
+  setHandlers(observers: readonly ClassType<Record<string, any>>[]): void {
     this.handlers = new Map(
       observers
         .map(observer => getQsoHandlers(observer))
@@ -138,9 +134,10 @@ export class QueuedSubscriptionObserverManager {
   ): Promise<void> {
     const handler = this.handlers.get(message.subscription_name);
     if (handler && handler.observer_class) {
-      const observer = new handler.observer_class(ctx);
+      // We use typedi to construct the resolver..
+      const observer = ctx.container.get(handler.observer_class);
       const hndlr = handler.handler.bind(observer);
-      return hndlr(message);
+      return hndlr(ctx, message);
     }
   }
 }